car-price-analyzer-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

sqlalchemy==1.4.51
databases[postgresql]==0.8.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
aiohttp==3.9.1
requests==2.31.0
lxml==4.9.3

numpy==1.24.4
pandas==2.0.3

python-dotenv==1.0.0
python-multipart==0.0.6
loguru==0.7.2

pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0